Rename app.js to index.js for more idomatic entrypoint naming
Generally people tend to use index.js or entry.js
